==> R/30_figures_main.R <==


if (!dir.exists("figures/main_text")){dir.create("figures/main_text", recursive = TRUE)}

# Figure 1
source("R/31_figure_1.R")

# Figure 2
source("R/32_figure_2.R")

# Figure 3
source("R/33_figure_3.R")

# Figure 4
source("R/34_figure_4.R")

# Figure 5
source("R/35_figure_5.R")
